// src/shared/hardware-profiles.steam.generated.ts
export const steamHardwareCatalog: SteamHardwareCatalog = {
  "windows": {
    "resolutions": [
      {
        "value": {
          "width": 1920,
          "height": 1080
        },
        "weight": 0.5415898381479205
      },
      {
        "value": {
          "width": 2560,
          "height": 1440
        },
        "weight": 0.22618315918869084
      },
      {
        "value": {
          "width": 2560,
          "height": 1600
        },
        "weight": 0.05511165744724442
      },
      {
        "value": {
          "width": 3840,
          "height": 2160
        },
        "weight": 0.05132145052243393
      },
      {
        "value": {
          "width": 3440,
          "height": 1440
        },
        "weight": 0.03206310182339685
      },
      {
        "value": {
          "width": 1920,
          "height": 1200
        },
        "weight": 0.027453390698627334
      },
      {
        "value": {
          "width": 1366,
          "height": 768
        },
        "weight": 0.025302192173734896
      },
      {
        "value": {
          "width": 1600,
          "height": 900
        },
        "weight": 0.007375537799631223
      },
      {
        "value": {
          "width": 2560,
          "height": 1080
        },
        "weight": 0.007170661749641466
      },
      {
        "value": {
          "width": 2880,
          "height": 1800
        },
        "weight": 0.005326777299733661
      },
      {
        "value": {
          "width": 1360,
          "height": 768
        },
        "weight": 0.004609711124769515
      },
      {
        "value": {
          "width": 5120,
          "height": 1440
        },
        "weight": 0.004199959024790001
      },
      {
        "value": {
          "width": 1440,
          "height": 900
        },
        "weight": 0.004097520999795125
      },
      {
        "value": {
          "width": 1680,
          "height": 1050
        },
        "weight": 0.0036877688998156115
      },
      {
        "value": {
          "width": 1280,
          "height": 720
        },
        "weight": 0.002356074574882196
      },
      {
        "value": {
          "width": 1280,
          "height": 1024
        },
        "weight": 0.00215119852489244
      }
    ],
    "cpuCores": [
      {
        "value": 6,
        "weight": 0.28432843284328435
      },
      {
        "value": 8,
        "weight": 0.27742774277427745
      },
      {
        "value": 4,
        "weight": 0.12451245124512453
      },
      {
        "value": 10,
        "weight": 0.06990699069906992
      },
      {
        "value": 16,
        "weight": 0.057705770577057704
      },
      {
        "value": 12,
        "weight": 0.05040504050405041
      },
      {
        "value": 14,
        "weight": 0.04920492049204921
      },
      {
        "value": 2,
        "weight": 0.030703070307030705
      },
      {
        "value": 24,
        "weight": 0.029102910291029106
      },
      {
        "value": 20,
        "weight": 0.021702170217021706
      },
      {
        "value": 5,
        "weight": 0.0020002000200020006
      },
      {
        "value": 18,
        "weight": 0.0015001500150015003
      },
      {
        "value": 3,
        "weight": 0.0009000900090009002
      },
      {
        "value": 1,
        "weight": 0.00020002000200020005
      },
      {
        "value": 7,
        "weight": 0.00010001000100010003
      },
      {
        "value": 22,
        "weight": 0.00010001000100010003
      },
      {
        "value": 28,
        "weight": 0.00010001000100010003
      },
      {
        "value": 32,
        "weight": 0.00010001000100010003
      }
    ],
    "ram": [
      {
        "value": 16,
        "weight": 0.41219290263876246
      },
      {
        "value": 32,
        "weight": 0.3844909513699323
      },
      {
        "value": 8,
        "weight": 0.07582650894752804
      },
      {
        "value": 64,
        "weight": 0.040238600748154885
      },
      {
        "value": 12,
        "weight": 0.0256799110302295
      },
      {
        "value": 24,
        "weight": 0.020018198362147403
      },
      {
        "value": 4,
        "weight": 0.015266403801435648
      },
      {
        "value": 48,
        "weight": 0.011930037407744412
      },
      {
        "value": 60,
        "weight": 0.009099181073703368
      },
      {
        "value": 28,
        "weight": 0.005257304620361944
      }
    ]
  },
// ...
};